poblacion inicial

+++ src/control/Estado.java
@@ -15,7 +15,7 @@ public class Estado {
 	
 	private float demanda;
 	private int numeroSeres;
-	private int dineroActual;
+	private float dineroActual;
 
 	public Estado(float demanda, int numeroSeres, int dineroActual) {
 		super();
@@ -68,7 +68,7 @@ public class Estado {
 		return numeroSeres;
 	}
 
-	public int getDineroActual() {
+	public float getDineroActual() {
 		return dineroActual;
 	}
